move sidebar button out of header
