+++ client/src/styled-components/HeaderStyles.js
@@ -1,11 +1,8 @@
 import {styled} from '@mui/material/styles';
-import {theme} from '../components/theme.jsx';
 import Box from '@mui/material/Box';
-import Button from '@mui/material/Button';
 
 export const HeaderLeftBox = styled(Box)(() => ({
-    display: 'flex',
-    alignItems: 'center',
+    marginLeft: '2.5em',
     fontSize: '1.5em', 
     fontWeight: 'bold',
     textAlign: 'center', 
@@ -16,12 +13,4 @@ export const HeaderRightBox = styled(Box)(() => ({
     width: 'fit-content',
     right: '1em',
     position: 'absolute'
-}));
-
-export const SidebarButton = styled(Button)(() => ({
-    zIndex: '1200',
-    fontSize: '25px',
-    fontWeight: '700',
-    color: '#888',
-    padding: '0'
 }));
